% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databutton.R
\name{Databutton}
\alias{Databutton}
\title{Databutton Client}
\description{
A class to interact with Databutton boards.
}
\details{
You can push and pull artifacts to and from the board.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Databutton$new()}}
\item \href{#method-authenticate}{\code{Databutton$authenticate()}}
\item \href{#method-get_data_frame}{\code{Databutton$get_data_frame()}}
\item \href{#method-push_data_frame}{\code{Databutton$push_data_frame()}}
\item \href{#method-clone}{\code{Databutton$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Databutton Client
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Databutton$new(token)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{token}}{Auth token - can be found in the board.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Databutton` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-authenticate"></a>}}
\if{latex}{\out{\hypertarget{method-authenticate}{}}}
\subsection{Method \code{authenticate()}}{
Authenticate with Databutton
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Databutton$authenticate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Authentification status
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data_frame"></a>}}
\if{latex}{\out{\hypertarget{method-get_data_frame}{}}}
\subsection{Method \code{get_data_frame()}}{
Pull a Dataframe from a Databutton board
Will download a table from a given Databutton board as data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Databutton$get_data_frame(board_id, table_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{board_id}}{The board ID, can be found when inspecting the table in Databutton}

\item{\code{table_id}}{The table ID, can be found when inspecting the table in Databutton}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data frame representation of the table
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-push_data_frame"></a>}}
\if{latex}{\out{\hypertarget{method-push_data_frame}{}}}
\subsection{Method \code{push_data_frame()}}{
Pull a Dataframe from a Databutton board
Will download a table from a given Databutton board as data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Databutton$push_data_frame(data_frame, board_id, table_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_frame}}{A Data Frame with the data to update the table_id with}

\item{\code{board_id}}{The board ID, can be found when inspecting the table in Databutton}

\item{\code{table_id}}{The table ID, can be found when inspecting the table in Databutton}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data frame representation of the table
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Databutton$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
